feat: moved the show-hide ts into a separate folder, run linting
