+++ src/ts/on-this-page/on-this-page.ts
@@ -1,4 +1,4 @@
-import initShowHideWithToggle from './show-hide';
+import initShowHideWithToggle from '../utils/show-hide';
 
 const initOnThisPage = (): void => {
   const onThisPageSelectors = {
