+++ src/ts/utils/tsconfig.json
@@ -1,0 +1,11 @@
+{
+  "extends": "../../../tsconfig-base.json",
+  "compilerOptions": {
+    "outDir": "../../../lib/",
+    "module": "CommonJS",
+    "rootDir": ".",
+    "declaration": true
+  },
+  "include": ["./**/*.ts"],
+  "exclude": ["./**/*.test.ts"]
+}
